spi: add driver for Axiado AX3000 SPI controller
Add the driver for the SPI Host controller on the Axiado AX3000 SoC and
Evaluation board.
This driver enables the kernel to control the SPI hardware and communicate
with peripheral devices on the bus.

The driver implements the core host functionalities, including:
- Probing based on the "axiado,ax3000-spi" compatible string.
- Interrupt-driven, FIFO-based transfers.
- Register access for controller configuration and status.

This patch depends on the dt-binding introduced in the prior commit.

Signed-off-by: Vladimir Moravcevic <vmoravcevic@axiado.com>
Signed-off-by: Swark Yang <syang@axiado.com>
Signed-off-by: Prasad Bolisetty <pbolisetty@axiado.com>
Signed-off-by: Tzu-Hao Wei <twei@axiado.com>