arm64: dts: axiado: enable SPI controller on AX3000
Add the device tree node for the SPI master controller to the Axiado
AX3000 SoC dtsi file.

This node defines the necessary properties such as the register base
address, interrupts, and clocks, according to the binding defined in an
earlier patch.

With this change, the spi-axiado driver can now match and probe the
device upon boot.

Signed-off-by: Swark Yang <syang@axiado.com>
Signed-off-by: Tzu-Hao Wei <twei@axiado.com>