dt-bindings: mfd: add Axiado SPI Bridge device
Add the device tree binding documentation for the Axiado SPI Bridge.

This component is a multi-function device (MFD) that acts as a bridge
between an SPI bus and its child devices. It provides a regmap interface
to these child nodes using a 3-phase SPI protocol.

The binding defines the compatible string "axiado,spi-bridge" and its
required properties.

Signed-off-by: Tzu-Hao Wei <twei@axiado.com>
Signed-off-by: Swark Yang <syang@axiado.com>