arm64: dts: axiado: add SPI bridge node for AX3000
Add the device tree node for the Axiado SPI bridge on the AX3000 SoC.

This node is placed under the main SPI controller and defines the chip
select and bus frequency. It acts as a parent node, allowing child
devices that are accessed via this bridge to be defined within it.

This enables the MFD driver to probe and provide services to the child
devices.

Signed-off-by: Tzu-Hao Wei <twei@axiado.com>
Signed-off-by: Swark Yang <syang@axiado.com>