soc: axiado: Add initial support and SGPIO driver
This patch introduces the initial platform support for Axiado SoCs.
It adds the necessary directory structure, Kconfig options, and Makefile
hooks under drivers/soc/axiado/.

As the first driver for this new platform, this commit adds the driver
for the Axiado SGPIO controller.

The Kconfig options SOC_AXIADO and AXIADO_SGPIO are introduced to
allow compile-time configuration of this support

Signed-off-by: Tzu-Hao Wei <twei@axiado.com>
Signed-off-by: Swark Yang <syang@axiado.com>