dt-bindings: i2c: cdns: Add compatible for regmap variant
Add the "cdns,i2c-r1p14-regmap" compatible string to the list of
supported devices for the Cadence I2C controller binding.

This compatible is intended for hardware variants where the I2C core
is part of a Multi-Function Device (MFD) and is accessed via a
shared regmap.

This is a preparatory patch for the new i2c-cadence-regmap driver,
which will be added in a subsequent commit.

Signed-off-by: Tzu-Hao Wei <twei@axiado.com>
Signed-off-by: Swark Yang <syang@axiado.com>