dt-bindings: gpio: Add binding for Axiado LTPI GPIO
Add a new YAML-based device tree binding for the Axiado LTPI GPIO
controller.

The LTPI GPIO controller functions as both a GPIO and an interrupt
controller. It provides a configurable number of dedicated output lines
and a corresponding number of input lines. The initial state of the
output lines can be specified in the device tree.

The controller supports both direct memory-mapped access and operation
through a parent device's regmap interface.

Signed-off-by: Tzu-Hao Wei <twei@axiado.com>
Signed-off-by: Swark Yang <syang@axiado.com>