MAINTAINERS: Add entry for Axiado LTPI controller
Add a new entry to the MAINTAINERS file specifically for the Axiado
LTPI controller driver.

This entry covers the core driver source files and the
corresponding device tree binding. Listing the maintainers ensures
that patches and questions related to this specific controller are
directed to the appropriate people for review.

Signed-off-by: Tzu-Hao Wei <twei@axiado.com>
Signed-off-by: Swark Yang <syang@axiado.com>