i3c: master: cdns: Add quirk to disable controller halt for AX3000
When an I3C hub (e.g., Renesas) that supports
IBIR handling is connected to the AX3000-i3c host
(which does not support IBIR), a controller halt
error can be triggered.

To prevent this halt state, add the
CDNS_I3C_QUIRKS_DISABLE_CTRL_HALT quirk to disable the
CTRL_HALT_EN (Halt on Error) feature for this controller.

Signed-off-by: Swark Yang <syang@axiado.com>
Signed-off-by: Tzu-Hao Wei <twei@axiado.com>