i3c: master: cdns: Fix possible divide-by-zero in thd_delay calculation
The sysclk rate is fetched using clk_get_rate(), which can
return 0 if the clock is not enabled or fails. This value
is later used as a divisor, creating a risk of a
divide-by-zero kernel panic.

Add a check for a zero sysclk_rate. If it is zero,
trigger a WARN_ON() for visibility and return a safe
(maximum encoded) delay value to prevent a crash.

Signed-off-by: Swark Yang <syang@axiado.com>
Signed-off-by: Tzu-Hao Wei <twei@axiado.com>