i3c: master: cdns: Fix potential sign extension in THD_DELAY logic
The cdns_i3c_master_calculate_thd_delay() function returns
a u8. Due to C's integer promotion rules, this value is
promoted to a signed 'int' before being passed to the
CTRL_THD_DELAY() macro.

If the value is large, this can lead to sign-extension
when the value is bit-shifted and OR'd into the u32 'ctrl'
register, corrupting other control bits.

Fix this by explicitly casting the return value to (u32)
before the macro is applied, ensuring no sign extension
can occur.

Signed-off-by: Swark Yang <syang@axiado.com>
Signed-off-by: Tzu-Hao Wei <twei@axiado.com>