arm64: dts: axiado: Set disable-ibir property for i3c nodes
Add the 'disable-ibir' property to the AX3000-i3c master
node in the ax3000.dtsi file.

The AX3000-i3c controller may generate spurious IBI
interrupts when standard I2C devices are connected to
the bus. Handling these unexpected IBIs can lead to a
driver crash.

Setting this property instructs the Cadence master driver
(which reads this property) to set a quirk flag that
disables IBIR handling entirely. This avoids the crash
on this platform.

Signed-off-by: Swark Yang <syang@axiado.com>
Signed-off-by: Tzu-Hao Wei <twei@axiado.com>